requirements.txt
numpy
opencv-python
numba
matplotlib
tqdm
vot-toolkit